backend/requirements.txt
fastAPI>=0.110.0
uvicorn>=0.15.0
pydantic>=1.9.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
requests>=2.26.0
sqlalchemy>=1.4.0
psycopg2>=2.9.1
psycopg2-binary>=2.9.1
pydantic>=1.9.0
pymupdf>=1.18.14
llama-index
llama-index-embeddings-huggingface